DjangoAPI/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.11.0
django-event-system==1.0.1
django-material==1.9.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.3.2
gevent==21.8.0
greenlet==1.1.2
idna==3.3
numpy==1.22.0
opencv-python==4.5.5.62
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.1
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
pyzbar==0.1.8
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.10.0
zope.event==4.5.0
zope.interface==5.4.0
